Add dynamicinfo.visibleinfo submodule
visibleinfo is a "watered-down" version of stateinfo and
contains only information that is accessible to a human player.
Along with the actual submodule, this commit adjusts the main
bot loop and the Agent code to accept both the full state and
the visible state. Pathfinding is also adjusted to allow for
tiles with unknown (nil) terrain codes.