Add held item equipping logic
Equip the best held item in the bag if possible (at appropriate times).
The data file item_priority.csv is now item_strategy.csv, and now
contains a separate priority for held items (used in this commit) and
a flag for whether an item should be used rather than thrown out (not
used now, but will be in future changes).