Fix stale heldBy property on forcibly dropped items
If an item previously held by a party member is forcibly dropped due to
a Trip Trap or Knock Off, the heldBy property will remain nonzero until
the item is picked up again by the player, even though it is no longer
held by anyone. This change forces items on the ground to have
heldBy == 0. In the case of a forcibly dropped item, a special droppedBy
property is set on the item with the index of the previous holder.