Use items rather than leaving them if it makes sense
With certain items, it's better to just use them if they would otherwise
be left on the ground (due to the bag having no space). For example, it
makes sense to eat berries and certain seeds, and use certain orbs,
since doing so can give minor benefits (like filling belly) that are
better than nothing.