Notice abnormal mobility of enemies
If an enemy is standing on a tile it wouldn't normally be able to
stand on, that should be taken as a sign that something is enhancing its
mobility, and pathfinding should take this into account. Otherwise, bad
things can happen in rare cases. For example, if the agent is trying to
rest on the stairs while next to an enemy, and that enemy is standing on
a tile it wouldn't normally be able to, the old agent (before this
change) would not recognize the enemy as nearby, and keep resting on the
stairs as the enemy kept attacking freely.