NFC: Create BaseAgent abstract base class for all Agent classes
Because code reuse and API standardization is nice. This change renames
BasicAgent to SimpleAgent, and has Agent and SimpleAgent inherit from a
base class called BaseAgent. BaseAgent contains the ugly boilerplate for
Lua classes, freeing the subclass files from needing it.

This change also moves the setupTurn/finalizeTurn/turnOngoing logic into
the BaseAgent:setupActFinalize() method so that the main execution loop
doesn't need to be messing with the bot state explicitly.