Don't heal if taking more damage than healed last turn
If facing a strong enemy, the current bot often got trapped in a cycle
of healing, only to take enough damage the next turn to put it in
healing range again. This burns through healing supplies without much
benefit. In these cases, the bot would likely be better served just
attacking rather than healing.

This change adds tracking for HP and whether or not a healing item was
used. If a healing item was just used last turn but current HP is lower
than what it was last turn, then it probably doesn't make sense to burn
another healing item.