Improve scanNearbyEntities
Instead of just searching the visibility region, which doesn't include
everything on screen if the dungeon is dark, now also search everything
on screen.

Furthermore, remove the default path size limit (allow arbitrary path
lengths by default), and instead explicitly exclude paths containing
positions going both out of visibility and off screen. The path size
limit was a hack to prevent undesirable circuitous paths to entities
that were only technically reachable, and explicitly excluding such
paths is more direct and doesn't impose an artificial limit on scanning
range.

These changes more fully reflect entities that are "nearby" enough
that we can detect them. In the case where an entity is on screen
position-wise but is not actually visible to the player, visibleinfo
will handle this and this entity will be excluded from the entities list
argument when passed in.