Act more aggressively if under attack from something dangerous
If at least 30 HP or 1/4 of total health was lost last turn, infer an
"under attack" status. When under attack, pay attention to more enemies
(even some off screen), be looser when classifying "close" enemies, skip
some preparatory actions if the enemy isn't close, and enable the use of
room-clearing moves when responding to assailants in order to attack
enemies (with moves) by any means necessary.

This is particularly useful for fighting in big rooms if foes have
room-clearing moves of their own, and are attacking from afar.