Handle target pathfinding failures more gracefully
If unable to find a path to the target, reset the target and try
pathfinding to a fresh target. This can happen if the target suddenly
becomes unreachable, such as when the player is suddenly warped to a
different part of the floor.

However, if pathfinding fails even with a fresh target, still throw an
error and exit, in order to surface the problem and prevent infinite
loops.