Adjust persistent status curing behavior
If the stairs are on screen, don't bother curing persistent statuses;
they'll be cured soon enough when climbing the stairs.

Also, from testing, the Decoy status isn't actually dangerous for the
player, so remove it from the lists of statuses to cure; teammates don't
seem to attack you even if you're a Decoy.