Fix how pathfinder.getPath handles avoided positions
The current implementation uses a special corner-cuttable terrain code
for all avoided positions. This isn't strictly correct since it will
make the pathfinder treat all avoided positions as corner-cuttable, even
if not all the original tiles were. For example, enemies in walls might
be marked as positions to avoid, and walls aren't corner-cuttable.

This change adds a second special terrain code (AVOID_CODE_HARD) for
avoided positions that are NOT corner-cuttable, and references the
original terrain's corner-cuttability when deciding which avoid code to
use for each avoided position.